Minor cleanup

+++ src/main/java/cs203t10/ryver/fts/account/Account.java
@@ -1,7 +1,5 @@
 package cs203t10.ryver.fts.account;
 
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 
 import javax.persistence.Entity;
@@ -11,15 +9,10 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.CascadeType;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 import cs203t10.ryver.fts.transaction.Transaction;
 
 
@@ -33,11 +26,11 @@ public class Account {
 
     @JsonProperty("id")
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long accountId;
+    private Integer accountId;
 
     @JsonProperty("customer_id")
     @NotNull(message = "Customer ID cannot be null")
-    private Long customerId;
+    private Integer customerId;
 
     @JsonProperty("balance")
     @NotNull(message = "Balance cannot be null")
